change: serviceListのコンポーネント化
